decided on solution for optional package import / plus refactoring
