+++ openfe/setup/__init__.py
@@ -11,7 +11,7 @@ from gufe import (
 
 from .atom_mapping import (LigandAtomMapping,
                            LomapAtomMapper, lomap_scorers,
-                           PersesAtomMapper, perses_mapper)
+                           PersesAtomMapper, perses_scorers)
 
 from .network import Network
 from . import ligand_network_planning
