+++ openfe/tests/setup/atom_mapping/test_perses_scorers.py
@@ -17,6 +17,7 @@ except ImportError:
 # TODO: test notimplemented error
 # TODO: clean test file
 
+
 def test_perses_regression(lomap_basic_test_files_dir,  # in a dir for lomap
                            lomap_basic_test_files):
